requirements.txt
dash==2.14.2
dash-bootstrap-components==1.5.0
pandas==2.1.4
plotly==5.18.0
numpy==1.26.2
mysql-connector-python==8.2.0
gunicorn==21.2.0
flask-cors
ollama 